feat: pagefind search
