+++ client/index.js
@@ -78,3 +78,16 @@ const _session = new Session();
 globalThis.PostDoc ??= _session
 
 export const session = _session;
+
+
+export async function navigateTo(url, replace) {
+    await globalThis.PostDoc.navigator.navigateTo(url, replace);
+}
+
+export function onLeave(callback, options) {
+    globalThis.PostDoc.navigator.registerOnLeaveCallback(callback, options);
+}
+
+export function onRender(callback, options) {
+    globalThis.PostDoc.navigator.registerOnRenderCallback(callback, options);
+}
