+++ lib/commands/build.js
@@ -7,6 +7,7 @@
  */
 
 import { build } from 'vite';
+import runSearchIndex from '../search-index.js';
 
 import Collector from '../collector.js';
 import PostDocCommand from '../command.js';
@@ -22,7 +23,6 @@ export default function createBuildCommand() {
 
         try {
           await collector.collectPages();
-
           await collector.writePages();
 
           const inlineConfig = await ViteConfiguration.createForProduction(
@@ -31,6 +31,7 @@ export default function createBuildCommand() {
 
           await build(inlineConfig);
         } finally {
+          await runSearchIndex();
           await collector.clear();
         }
       })
