+++ lib/commands/preview.js
@@ -12,6 +12,7 @@
  */
 
 import { build, preview } from 'vite';
+import runSearchIndex from '../search-index.js';
 
 import Logger from '../logger.js';
 import Collector from '../collector.js';
@@ -35,6 +36,7 @@ export const runPreview = async ({ host, open, skipBuild = false } = {}) => {
     } catch {
       return;
     } finally {
+      await runSearchIndex();
       await collector.clear();
     }
   }
@@ -52,11 +54,10 @@ export const runPreview = async ({ host, open, skipBuild = false } = {}) => {
     () => `
       Server is listening on:
         - ${server.resolvedUrls.local.join(' | ')}
-        ${
-  server.resolvedUrls.network.length
-    ? '- ' + server.resolvedUrls.network.join(' | ')
-    : ''
-}
+        ${server.resolvedUrls.network.length
+        ? '- ' + server.resolvedUrls.network.join(' | ')
+        : ''
+      }
     `,
     Logger.SuccessLevel
   );
